Refactor basic modal for reusability
