+++ frontend/src/components/baseComponents/FundProposalButton.js
@@ -1,4 +1,6 @@
+import { useState } from "react";
 import BasicModal from "./BasicModal";
+import { ContributeForm } from "./ContributeForm";
 
 export const FundProposalButton = ({
   proposalAddress,
@@ -6,6 +8,9 @@ export const FundProposalButton = ({
   fundingStatus,
   providerStatus,
 }) => {
+  const [open, setOpen] = useState(false);
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
   if (
     fundingStatus === "Accepting Contributions" &&
     providerStatus === "Awaiting Provider"
@@ -15,6 +20,15 @@ export const FundProposalButton = ({
         sx={{ marginLeft: "0px" }}
         buttonTitle="Contribute"
         modalTitle="How much would you like to contribute to this campaign?"
+        open={open}
+        handleClose={handleClose}
+        handleOpen={handleOpen}
+        modalBody={
+          <ContributeForm
+            handleClose={handleClose}
+            amountRemaining={amountRemaining}
+          />
+        }
         amountRemaining={amountRemaining}
       />
     );
